+++ frontend/src/index.tsx
@@ -1,70 +1,14 @@
-import React from 'react';
-import type { NextPage } from 'next';
-import Link from 'next/link';
-import { Button, Container, Form, Col, Row } from 'react-bootstrap';
-import { useState } from 'react';
-import { useRouter } from 'next/router';
-import { callServer } from '../components/lib/apiCom';
+import * as React from "react";
+import * as ReactDOM from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
 
-const Login: NextPage = () => {
-	const router = useRouter();
-	const [email, setEmail] = useState('');
-	const [passcode, setCode] = useState('');
+import "./index.css";
+import App from "./App";
 
-	async function submitLogin() {
-		const loginInfo = { email: email, password: passcode };
-		callServer('/login', 'POST', loginInfo)
-			.then((response) => {
-				if (!response.ok) {
-					throw response;
-				}
-				return response.json();
-			})
-			.then((json) => {
-				localStorage.setItem(`essentialMailToken`, JSON.stringify(json));
-				router.push(`/mail/home`);
-			})
-			.catch((err) => {
-				alert(`Error logging in, please try again.`);
-			});
-	}
-
-	return (
-		<main className="background">
-			<Container className="centerpiece">
-				<Col>
-					<Form>
-						<Form.Group>
-							<Form.Label>Email Address</Form.Label>
-							<Form.Control
-								type="email"
-								placeholder="Email"
-								value={email}
-								onChange={(e) => setEmail(e.target.value)}
-							/>
-						</Form.Group>
-						<Form.Group className="mb-3" controlId="formBasicPassword">
-							<Form.Label>Password</Form.Label>
-							<Form.Control
-								type="password"
-								placeholder="Password"
-								value={passcode}
-								onChange={(e) => setCode(e.target.value)}
-							/>
-						</Form.Group>
-					</Form>
-					<Col>
-						<Button variant="primary" type="submit" onClick={submitLogin}>
-							Submit
-						</Button>
-						<div>
-							<Link href="/register">Register a new account</Link>
-						</div>
-					</Col>
-				</Col>
-			</Container>
-		</main>
-	);
-};
-
-export default Login;
+ReactDOM.createRoot(document.getElementById("root")!).render(
+  <React.StrictMode>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </React.StrictMode>
+);
